requirements.txt
flask
flask-login
tensorflow
numpy
pillow
python-dotenv
google-auth-oauthlib
requests
oauthlib
gunicorn